+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- Tax summary: Gains in the year should exclude losses
 
 ## Version [0.2.0] Beta (2019-10-30)
 ### Fixed